+++ src/instructions/micro/jump.rs
@@ -7,7 +7,7 @@ use crate::state::{Flags, Register16, State};
 ///
 /// This can be used to implement any kind of branch with a one byte immediate operand by adding
 /// the rest of the logic.
-pub fn jr_d(state: &mut State, cycles: u8) -> ExecResult {
+pub fn jr_d(state: &mut State, cycles: u32) -> ExecResult {
     let pc = state.pc().wrapping_add(state.z() as i8 as i16 as u16);
     *state.pc_mut() = pc.to_le_bytes();
     ExecResult::Done(cycles)
@@ -16,7 +16,7 @@ pub fn jr_d(state: &mut State, cycles: u8) -> ExecResult {
 /// Decrement B and jump to immediate offset if not zero.
 ///
 /// The number of cycles depends on if the instruction jumps or not.
-pub fn djnz_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecResult {
+pub fn djnz_d(state: &mut State, cycles_jump: u32, cycles_not_jump: u32) -> ExecResult {
     let b = state.b().wrapping_sub(1);
     *state.b_mut() = b;
     if b != 0 {
@@ -27,7 +27,7 @@ pub fn djnz_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecRe
 }
 
 /// Jump if the zero flag isn't set
-pub fn jr_nz_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecResult {
+pub fn jr_nz_d(state: &mut State, cycles_jump: u32, cycles_not_jump: u32) -> ExecResult {
     if state.get_flags().is_set(Flags::Z) {
         ExecResult::Done(cycles_not_jump)
     } else {
@@ -36,7 +36,7 @@ pub fn jr_nz_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecR
 }
 
 /// Jump if the zero flag is set
-pub fn jr_z_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecResult {
+pub fn jr_z_d(state: &mut State, cycles_jump: u32, cycles_not_jump: u32) -> ExecResult {
     if state.get_flags().is_set(Flags::Z) {
         jr_d(state, cycles_jump)
     } else {
@@ -45,7 +45,7 @@ pub fn jr_z_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecRe
 }
 
 /// Jump if the carry flag isn't set
-pub fn jr_nc_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecResult {
+pub fn jr_nc_d(state: &mut State, cycles_jump: u32, cycles_not_jump: u32) -> ExecResult {
     if state.get_flags().is_set(Flags::C) {
         ExecResult::Done(cycles_not_jump)
     } else {
@@ -54,7 +54,7 @@ pub fn jr_nc_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecR
 }
 
 /// Jump if the carry flag is set
-pub fn jr_c_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecResult {
+pub fn jr_c_d(state: &mut State, cycles_jump: u32, cycles_not_jump: u32) -> ExecResult {
     if state.get_flags().is_set(Flags::C) {
         jr_d(state, cycles_jump)
     } else {
@@ -65,7 +65,7 @@ pub fn jr_c_d(state: &mut State, cycles_jump: u8, cycles_not_jump: u8) -> ExecRe
 /// Pops the stack and jump to the value loaded into `WZ`.
 ///
 /// This assumes the value pointed by `SP` was loaded into `WZ`
-pub fn ret(state: &mut State, cycles: u8) -> ExecResult {
+pub fn ret(state: &mut State, cycles: u32) -> ExecResult {
     // Jump
     *state.pc_mut() = state.wz_bytes();
     // Pop
@@ -93,7 +93,7 @@ pub fn push(state: &mut State, reg: Register16) -> ExecResult {
 }
 
 /// Jump to the immediate value if the condition is true
-pub fn jp_cc_nn(state: &mut State, cond: bool, cycles: u8) -> ExecResult {
+pub fn jp_cc_nn(state: &mut State, cond: bool, cycles: u32) -> ExecResult {
     if cond {
         *state.pc_mut() = state.wz_bytes();
     }
@@ -101,7 +101,7 @@ pub fn jp_cc_nn(state: &mut State, cond: bool, cycles: u8) -> ExecResult {
 }
 
 /// Push `PC` into the stack if the condition is true. Otherwise, finish instruction execution
-pub fn push_pc_or_break(state: &mut State, cond: bool, cycles: u8) -> ExecResult {
+pub fn push_pc_or_break(state: &mut State, cond: bool, cycles: u32) -> ExecResult {
     if cond {
         let sp = state.sp().wrapping_sub(2);
         *state.sp_mut() = sp.to_le_bytes();
@@ -116,13 +116,13 @@ pub fn push_pc_or_break(state: &mut State, cond: bool, cycles: u8) -> ExecResult
 }
 
 /// Jump to the immediate value
-pub fn jr_mm(state: &mut State, cycles: u8) -> ExecResult {
+pub fn jr_mm(state: &mut State, cycles: u32) -> ExecResult {
     *state.pc_mut() = state.wz_bytes();
     ExecResult::Done(cycles)
 }
 
 /// Jump to the address
-pub fn jp(state: &mut State, address: u16, cycles: u8) -> ExecResult {
+pub fn jp(state: &mut State, address: u16, cycles: u32) -> ExecResult {
     *state.pc_mut() = address.to_le_bytes();
     ExecResult::Done(cycles)
 }
