+++ src/instructions/micro/ld.rs
@@ -6,7 +6,7 @@ use crate::state::{Register, Register16, State};
 /// Load an immediate value into a 16 bits register
 ///
 /// Returns `Done`
-pub fn ld_rr_nn(state: &mut State, register: Register16, cycles: u8) -> ExecResult {
+pub fn ld_rr_nn(state: &mut State, register: Register16, cycles: u32) -> ExecResult {
     state.set_register_16(register, state.wz());
     ExecResult::Done(cycles)
 }
@@ -24,7 +24,7 @@ pub fn ld_pp_r(state: &State, reg16: Register16, reg8: Register) -> ExecResult {
 /// Load an immediate value to an 8-bit register
 ///
 /// Returns `Done`
-pub fn ld_r_n(state: &mut State, register: Register, cycles: u8) -> ExecResult {
+pub fn ld_r_n(state: &mut State, register: Register, cycles: u32) -> ExecResult {
     state.set_register_8(register, state.z());
     ExecResult::Done(cycles)
 }
@@ -48,13 +48,13 @@ pub fn ld_mm_r(state: &State, reg: Register) -> ExecResult {
 }
 
 /// Load the contents of a register into another
-pub fn ld_r_r(state: &mut State, dst: Register, src: Register, cycles: u8) -> ExecResult {
+pub fn ld_r_r(state: &mut State, dst: Register, src: Register, cycles: u32) -> ExecResult {
     state.set_register_8(dst, state.get_register_8(src));
     ExecResult::Done(cycles)
 }
 
 /// Load the contents of a 16-bit register into another
-pub fn ld_rr_rr(state: &mut State, dst: Register16, src: Register16, cycles: u8) -> ExecResult {
+pub fn ld_rr_rr(state: &mut State, dst: Register16, src: Register16, cycles: u32) -> ExecResult {
     state.set_register_16(dst, state.get_register_16(src));
     ExecResult::Done(cycles)
 }
